fix download scheme with HUNTER_DOWNLOAD_SERVER
HUNTER_DOWNLOAD_SERVER replaces the single download URL with multiple
ones to be tried.
The download scheme uses file(DOWNLOAD), which does not support a list
of ULRs. When used together this causes an error

As a side effect of this commit the download scheme now supports
HUNTER_PACKAGE_PROTECTED_SOURCES.

fixes: https://github.com/cpp-pm/hunter/issues/79