szip: Fix example compile with size_t and extern "C"
szlib is compiled as a pure C static library. So there is no C++
name-mangling done. When including the `szlib.h` header into a C++
project the compiler assumes the header to be a C++ header and does
name-mangling. This results in linker errors when using szip-functions.

The second problem is the usage of `size_t` inside `szlib.h` whithout a
`stddef.h` include. Probably some other include provided the `size_t` in
the past, but maybe the stdlib changed and this missing include is
surfacing now.