docs: move to `html_css_files` as `add_stylesheet` is deprecated/removed
With Sphinx 4 the function `add_stylesheet()` was temporarily removed
and added back in in newer versions of Sphinx. As alternative the
`html_css_files` variable is mentioned, which is available since
Sphinx 1.8.

The `python3-sphinx` package with version `4.3.2-1` from Ubuntu 22.04
jammy doesn't have the `add_stylesheet()` API, which breaks
documentation generation on said platform.

https://github.com/sphinx-doc/sphinx/issues/9683#issuecomment-932878909