hunter_setup_msvc: tolower to fix arm64 vs ARM64 confusion
`HUNTER_MSVC_ARCH` should match the capitalization of the `vcvarsall.bat
<arch>` argument. The arch argument according to `vcvarsall.bat /help`
can be one of the following:

```
[arch]: x86 | amd64 | x86_amd64 | x86_arm | x86_arm64 | amd64_x86 | amd64_arm | amd64_arm64
```

But to be more like `vcvarsall.at` and `cmake -A`, be case-insensitive
with the input architectures.

And for less confusion, always output the `HUNTER_ARCH` variables in
lower case.

In the process cleanup, fix and extend the corresponding unittest.

Fixes: https://github.com/cpp-pm/hunter/issues/679